some more

+++ ring_buf.c
@@ -2,6 +2,7 @@
 #include <stdlib.h>
 #include <stdbool.h>
 #include "ring_buf.h"
+#include <time.h>
 
 RingBuf rb_init(size_t capacity){
     RingBuf rb;
@@ -79,7 +80,7 @@ float rb_avg(RingBuf *rb){
     float sum = 0;
     float avg;
 
-    for (int i = 0; i != rb->count; i++){   
+    for (size_t i = 0; i != rb->count; i++){   
 
         sum += rb->buf[(rb->tail + i) % rb->capacity].value;
     }
